Refactor chord explorer mode into separate class with ChordModeBase
